* FIX [mqtt_client] fix compile warning

+++ src/mqtt/protocol/mqtt/mqtt_client.c
@@ -290,7 +290,7 @@ mqtt_sock_set_retry_wait(void *arg, const void *v, size_t sz, nni_opt_type t)
 	return (rv);
 }
 
-static uint64_t
+static int
 mqtt_sock_get_pipeid(void *arg, void *buf, size_t *szp, nni_type t)
 {
 	// For MQTT Client, only has one pipe
